Separate prepare_data into functions and main
